+++ CHANGELOG.md
@@ -6,6 +6,9 @@ rename `Unreleased` topic with the new version tag. Finally, create a new `Unrel
 
 ## Unreleased
 
+### API: window
+- Export the `Neutralino.window.print()` function.
+
 ## v6.1.0
 
 ### API: window